IPACM: Fix to clean up default LL routes
When IPv4 only call is brought up, interface will still have v6
LL address and we go ahead and add default routes. When the call
is torn down, we do not clean this as there is no global v6 address.
Make a change to clean default LL routes in this case.

Change-Id: I7e203aa82a03351cb94f65ea89fa57bd3f1059a9