ipacm: Fix to ignore stale private ip in passthrough mode
When IP passthrough mode is activated, the neighbour entry
for the private ip will not be deleted immediately and we
get NEWNEIGH events from kernel. As a result we add incorrect
routing rules towards the client ep. Make a change to ignore
the stale private ip if the LAN client is in IP passthrough
mode.

Change-Id: Ia02fc109e13f324a3f740642718f59f5ee54c160