ipacm: Add support IP Passthrough w/o NAT
Make changes to get IP Passthrough configuration
via ioctl and add support to add alg entries to IPA HW
when NAT needs to be skipped. Even though this is w/o NAT
we still add NAT entries to HW which will be dummy operation.

Also fix a legacy issue to not allow ALG entries to be added
as dummy NAT entries as packets will be incorrectly routed
to Passthrough client.

Change-Id: If35b1c41305d68fb2316769ce55fd59ccf960fe9