ipacm: Add variable to store no of q6 firewall rules.
While configuring the number of UL firewall rules, to
be sent to Q6 side, no of firewall rules configured
was not cached which was resulting in using incorrect
no of firewall rules sent to Q6. Now adding new variable
to cache no of ul firewall rule configured to be send to
Q6 side.

Change-Id: Iedcf83f97d0c49c32d215999653cfe2d497958ac