ipacm: Add handling for vlan bridge down/up event
Update the mac addr for the same vlan bridge when it goes
down and up. Add a new API to get bridge from vlan id.
Use this API to get the bridge directly in handle vlan
neighbor to prevent corruption of the variable from
typecasting it to a larger struct.

Change-Id: Id6576ef82bc62afefaa15ec146a53b07324c9e83