ipacm: Fix to handle nat table switch failure while table is switching between SRAM to DRR multiple times
issue: While performing speed test multiple conntracks are
established and if they remain stable for a long time without
getting destroyed and table is switching between SRAM and DRR
multiple times in this course of time, it will lead to nat
table switching failure. Here the cath is nat rule for the
new conntrack is added properly to the HW but will fail to
add to the maps which we use for table transition between
SRAM and DDR.

Added fix to get a new entry value while adding entry to maps
without disturbing the current rule hdl of SRAM/DRR table.
This new handle will be returned to ipacm which we use later
for table transition between SRAM and DDR.

Change-Id: I7feed4461fd8a6c0b377ff59a87bca86a3a7ba45
(cherry picked from commit 619383e9dc56cc4ce71dcad4586e3161a4afb87d)