data-ipa-cfg-mgr: Resolves makefile violations in ipacm
During the makefile violation scan for the Bazel migration on Android-V,
cnt_macro_errors were identified in the ipacm makefile. The current is-xxx macros have several issues, as they are awkwardly used by comparing their result with true instead of the conventional method of checking if the value is not empty. Therefore, these macros are being replaced with a new, more convenient set of macros defined in build/make/core.

Change-Id: I7edb7d4a1745abc787a3b8812a3f9455ce0523dd
Signed-off-by: Jagadeesh Ponduru <quic_jponduru@quicinc.com>