ipacm: Fix to post eogre events only on ethernet ifaces
Issue: When USB composition is changes from 90DB to 902D, DL data
stall is observed for EoGRE traffic. When ECM Connect is received
for RnDIS, IPACM processes this as a EoGRE event and updates the DL routing rules permanently with USB Destination pipe from ETH.

Added iface check while we post eogre_down and eogre_up so that we
post these events only for eth iface.

Change-Id: I445d77a9f6c50460471af1aee93aee6394bad865