ipacm: Fix stats support when siface/vlan is enabled
When LAN stats are enabled for an Ethernet interface configured as VLAN,
IPACM was not installing uplink filter rules on non-VLAN pipes for
siface/VLAN configured interfaces. This caused packets to take filter
exceptions.

Add support to install filter rules for non-VLAN siface when an interface is configured as VLAN. This ensures proper packet handling
and statistics collection for these configurations.

Change-Id: Ieccb8edc277c851f829004af9739943eee608288